refactor(core): Remove ChangeDetectionScheduler abstract class
The class was added at some point both for configurability of the scheduler during
its early days of implementation and I think also for cyclic imports at some point(?).

Neither of those are necessary anymore. This simplifies the internal abstractions.
